feat(web): replace PyQt6 desktop GUI with Streamlit web app
Move the front end to Streamlit, mirroring the WrinkleFE layout: sidebar
inputs feed a cached run_analysis function whose result fans out to
Profile, Mesh, Results, Stress, and Export tabs. The pure helpers
(parse_layup, build_export_payload, write_results_json/csv) are kept
module-level on app.py so existing tests import them unchanged.

- Add app.py (Streamlit), DEPLOYMENT_STREAMLIT.md, requirements-test.txt
- requirements.txt: drop PyQt6, add streamlit; bump matplotlib to 3.9+
- pyproject.toml: replace gui extras with web extras, drop porosity-fe
  console script and porosity_gui from py-modules
- Delete porosity_gui.py and PorosityFE.spec; keep ValidatePorosity.spec
  (CLI is unaffected)
- tests workflow: install from requirements*.txt and smoke-test app
  imports instead of porosity_gui
- README: document streamlit run app.py and link DEPLOYMENT_STREAMLIT.md
