+++ requirements.txt
@@ -1,5 +1,5 @@
 numpy>=1.20.0
 scipy>=1.7.0
-matplotlib>=3.4.0
+matplotlib>=3.9,<4
 jsonschema>=4.0.0
-PyQt6>=6.5.0
+streamlit>=1.32
